--- a/src/librepgp/stream-common.c
+++ b/src/librepgp/stream-common.c
@@ -324,7 +324,7 @@ dst_write(pgp_dest_t *dst, const void *buf, size_t len)
         /* if cache non-empty and len will overflow it then fill it and write out */
         if ((dst->clen > 0) && (dst->clen + len > sizeof(dst->cache))) {
             memcpy(dst->cache + dst->clen, buf, sizeof(dst->cache) - dst->clen);
-            buf = (uint8_t*)buf + sizeof(dst->cache) - dst->clen;
+            buf = (uint8_t *) buf + sizeof(dst->cache) - dst->clen;
             len -= sizeof(dst->cache) - dst->clen;
             dst->werr = dst->write(dst, dst->cache, sizeof(dst->cache));
             dst->writeb += sizeof(dst->cache);