This PR fixes some performance bottlenecks in RNP:
- adds table-based crc-24 calculation to armouring/dearmoring streams
- fixes slow cfb encryption/decryption
- adds caching to stream writes
- updates performance tests

Within this changes for main encryption algorithms (AES, Camellia) RNP is 1.7-2 times faster then GnuPG (on my machine). Some ciphers, like Blowfish, Cast5 are slower. I didn't dig deep, but it seems that for those Botan has slower implementation then for AES/Camellia. @randombit, what do you think?

Most likely there are more things which may be improved. With these changes performance was improved in 1.5-2.0 times.

The results are as following:
```
#1. Small file symmetric encryption
ENCRYPT-SMALL-BINARY:RNP:26.22 runs/sec
ENCRYPT-SMALL-BINARY:GPG:39.44 runs/sec
ENCRYPT-SMALL-BINARY:RNP vs GPG:1.50
ENCRYPT-SMALL-ARMOUR:RNP:25.43 runs/sec
ENCRYPT-SMALL-ARMOUR:GPG:40.06 runs/sec
ENCRYPT-SMALL-ARMOUR:RNP vs GPG:1.58
#2. Large file symmetric encryption
ENCRYPT-AES128-BINARY:RNP:154.76 MB/sec
ENCRYPT-AES128-BINARY:GPG:70.71 MB/sec
ENCRYPT-AES128-BINARY:RNP vs GPG:0.46
ENCRYPT-AES192-BINARY:RNP:125.19 MB/sec
ENCRYPT-AES192-BINARY:GPG:62.10 MB/sec
ENCRYPT-AES192-BINARY:RNP vs GPG:0.50
ENCRYPT-AES256-BINARY:RNP:144.64 MB/sec
ENCRYPT-AES256-BINARY:GPG:71.48 MB/sec
ENCRYPT-AES256-BINARY:RNP vs GPG:0.49
ENCRYPT-TWOFISH-BINARY:RNP:86.25 MB/sec
ENCRYPT-TWOFISH-BINARY:GPG:52.35 MB/sec
ENCRYPT-TWOFISH-BINARY:RNP vs GPG:0.61
ENCRYPT-BLOWFISH-BINARY:RNP:59.89 MB/sec
ENCRYPT-BLOWFISH-BINARY:GPG:75.72 MB/sec
ENCRYPT-BLOWFISH-BINARY:RNP vs GPG:1.26
ENCRYPT-CAST5-BINARY:RNP:58.39 MB/sec
ENCRYPT-CAST5-BINARY:GPG:69.47 MB/sec
ENCRYPT-CAST5-BINARY:RNP vs GPG:1.19
ENCRYPT-CAMELLIA128-BINARY:RNP:83.33 MB/sec
ENCRYPT-CAMELLIA128-BINARY:GPG:50.05 MB/sec
ENCRYPT-CAMELLIA128-BINARY:RNP vs GPG:0.60
ENCRYPT-CAMELLIA192-BINARY:RNP:68.03 MB/sec
ENCRYPT-CAMELLIA192-BINARY:GPG:42.69 MB/sec
ENCRYPT-CAMELLIA192-BINARY:RNP vs GPG:0.63
ENCRYPT-CAMELLIA256-BINARY:RNP:70.08 MB/sec
ENCRYPT-CAMELLIA256-BINARY:GPG:44.24 MB/sec
ENCRYPT-CAMELLIA256-BINARY:RNP vs GPG:0.63
#3. Large file armoured encryption
ENCRYPT-LARGE-ARMOUR:RNP:64.93 MB/sec
ENCRYPT-LARGE-ARMOUR:GPG:43.85 MB/sec
ENCRYPT-LARGE-ARMOUR:RNP vs GPG:0.68
```