@ronaldtse 2017-10-09
@ni4 this is impressive stuff.

For the CLI test output multiple, could we write "XX% faster" or "XX% slower" rather than "0.63"

I assume the `ENCRYPT_{SMALL,LARGE}_` tests are all using AES CFB? I wonder why `rnp` is so much slower with small files when it's so much faster with large ones.

It feels either something to do with the stream setup or memory allocation (and/or the finalization of it). Maybe if we can graph out memory allocation comparison we can find out why.