@flowher 2017-10-09
Does your gnupg support hardware instructions?