@randombit 2017-10-10
Looks good to me though I didn't read the stream cache logic too carefully.

One thing, did you try seeing what gcc/clang do if you just use `uint64_t` everywhere for the CFB code? I would think GCC would know how to convert this into a pair of 32-bit registers with basically no extra overhead. Which would simplify the code a little bit.

BTW, some caution here, I did something very much like what you did for CFB XOR at one point in Botan and ran into problems because GCC -O3 will vectorize the XORs. The problem is that when GCC vectorized it, it used aligned SSE loads (even though you can normally do misaligned loads on x86 just fine), so the code would crash if fed a misaligned buffer. Maybe this is fixed in recent GCC, don't know.

Re Blowfish and CAST perf, I'm not sure why these would be particularly slower than what libgcrypt is doing... however did a little digging and checked on asm dumps, for CAST I see GCC is doing something kind of non-optimal, because we have to be careful to avoid undefined behavior in the bit rotation function, but GCC misses the fact that one of our operations is redundant. This caused ~15% slowdown (in the overall CAST-128 encryption) vs just doing the "naive" rotation instrinsic. I filed a bug with GCC https://gcc.gnu.org/bugzilla/show_bug.cgi?id=82498. I briefly played around with trying to get something that both avoids undefined behavior and gets GCC to generate decent code but no luck so far.

Actually also it appears Clang generates terrible code for `rotate_left` in Botan, doesn't even generate the `roll` instruction instead does shift/shift/or :(