@randombit 2017-10-10
Re the performance of CAST-128 I improved things ~20% by using asm for the rotation - https://github.com/randombit/botan/pull/1247

Don't know how that will affect the overall rnp performance but it's something.