@ni4 2017-10-10
@flowher Reasonable comment, thanks. I used older GnuPG installation with libgcrypt 1.7.7. Now I updated it to 1.8.1 and speeds are comparable to RNP - 150Mb/sec for AES. However, homebrew formula is built with '--disable-asm'. Will investigate it a little bit more.

@ronaldtse Sure, I'll update output of the performance test for better readability, it is awkward now.