@ni4 2017-10-10
@randombit Thanks for the detailed comment. Sure, will update to not use #ifdefs, just uint64_t. Also got warnings about unaligned 64-bit pointer access, will investigate it. As I quick-read yesterday unaligned access may slow the things down twice. So needs some more investigation.