@flowher 2017-10-10
>  Also got warnings about unaligned 64-bit pointer access, will investigate it
Problems, problems, problems. I've been there and wasted a lot of time debugging. Better fix it before commit