@ni4 2017-10-11
@ronaldtse I updated the output format, please see numbers for GnuPG with libgcrypt 1.8.1 below
Also rebased/removed #ifdefs as @randombit suggested.

Will investigate slowdown on runs/sec and problems with unaligned memory access later today.

```
#1. Small file symmetric encryption
ENCRYPT-SMALL-BINARY          : RNP is  55% SLOWER then GnuPG (28.01 runs/sec vs 61.70 runs/sec)
ENCRYPT-SMALL-ARMOUR          : RNP is  54% SLOWER then GnuPG (28.23 runs/sec vs 61.36 runs/sec)
#2. Large file symmetric encryption
ENCRYPT-AES128-BINARY         : RNP is   3% SLOWER then GnuPG (170.28 MB/sec vs 175.38 MB/sec)
ENCRYPT-AES192-BINARY         : RNP is   1% SLOWER then GnuPG (166.55 MB/sec vs 169.05 MB/sec)
ENCRYPT-AES256-BINARY         : RNP is   1% SLOWER then GnuPG (156.39 MB/sec vs 158.44 MB/sec)
ENCRYPT-TWOFISH-BINARY        : RNP is   2% SLOWER then GnuPG (89.61 MB/sec vs 91.19 MB/sec)
ENCRYPT-BLOWFISH-BINARY       : RNP is  20% SLOWER then GnuPG (59.27 MB/sec vs 74.32 MB/sec)
ENCRYPT-CAST5-BINARY          : RNP is  18% SLOWER then GnuPG (59.71 MB/sec vs 72.48 MB/sec)
ENCRYPT-CAMELLIA128-BINARY    : RNP is   7% FASTER then GnuPG (87.01 MB/sec vs 81.63 MB/sec)
ENCRYPT-CAMELLIA192-BINARY    : RNP is   6% FASTER then GnuPG (73.04 MB/sec vs 68.67 MB/sec)
ENCRYPT-CAMELLIA256-BINARY    : RNP is   8% FASTER then GnuPG (74.05 MB/sec vs 68.37 MB/sec)
#3. Large file armoured encryption
ENCRYPT-LARGE-ARMOUR          : RNP is   8% FASTER then GnuPG (68.75 MB/sec vs 63.77 MB/sec)
```