@ronaldtse 2017-10-11
@ni4 somehow Travis is still failing on this build. Looks good to merge once this is fixed!

PS. I wonder about the differences between this run and @flowher 's run in #492 ?