@ni4 2017-10-11
@ronaldtse Build fails because of the alignment error in sanitize run: `SUMMARY: AddressSanitizer: undefined-behavior symmetric.c:307:17 in 
symmetric.c:309:23: runtime error: load of misaligned address 0x62d000000416 for type 'uint64_t' (aka 'unsigned long'), which requires 8 byte alignment`. 

So, trying to find a good way to overcome this.