@ronaldtse 2017-10-13
@randombit @dewyatt @flowher any quick solutions to the alignment issue? Would like to merge this ASAP. Thanks!