@ni4 2017-10-13
@ronaldtse I only found a way to use memcpy to temporary aligned buffer, if parameters are misaligned. And make cache in encrypted stream, which is used most times, aligned. Will finish these changes today or tomorrow.