@ni4 2017-10-14
@ronaldtse Alignment issue is now fixed. I will further investigate why we are so slow on small files, and check tests. Currently encryption itself doesnt' take much time, according to profiler. See the attached screenshot:

<img width="697" alt="screen shot 2017-10-14 at 1 50 38 pm" src="https://user-images.githubusercontent.com/4312096/31575080-ffb8cff6-b0e6-11e7-88b9-ff678a53bf40.png">