@ni4 2017-10-14
@randombit Btw, why spending in botan_block_cipher_encrypt_blocks around 1.10s, only 600ms are spent inside of AES encryption itself?