@ronaldtse 2017-10-15
Any further comments or can we merge this first?