@flowher 2017-10-15
I think I'm OK. Let me satar at it few mins more.